Restore TLS handling and warmup
