--- templates/chord.html
@@ -1,18 +1,0 @@
-{message_html}
-<div class="container">
-  <h2>Ableton Move Chord Kit Generator</h2>
-  <p>Select a .WAV, .AIF, .AIFF or .MP3 file to generate chords for a Choke Kit ablpreset.</p>
-  <input type="file" id="wavFileInput" accept=".wav, .aiff, .aif, .mp3"><br>
-  <label for="presetName">Preset Name (optional): </label>
-  <input type="text" id="presetName" placeholder="Preset name"><br>
-  <button id="generatePreset">Download .ablpresetbundle</button>
-  <button id="placePreset">Save Preset directly on Move</button>
-  <div id="loadingIndicator" style="display:none; margin-top:10px;">
-    Generating bundle... <span id="progressPercent">0%</span>
-  </div>
-  
-  <div class="chord-list">
-    <h3>Chords to be generated:</h3>
-    <ul id="chordList"></ul>
-  </div>
-</div>
