--- templates/index.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Move - Extended</title>
-        <link rel="stylesheet" href="/static/style.css">
-        
-        <!-- Include WaveSurfer core and the Regions plugin -->
-        <script src="https://unpkg.com/wavesurfer.js@6/dist/wavesurfer.js"></script>
-        <script src="https://unpkg.com/wavesurfer.js@6/dist/plugin/wavesurfer.regions.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/jszip/3.7.1/jszip.min.js"></script>
-        <script src="/static/main.js"></script>
-        <script src="/static/shared.js"></script>
-        <script src="/static/chord.js"></script>
-
-    </head>
-    <body>
-        <h1 id="header">Move - Extended</h1>
-        <div class="tab">
-            <button class="tablinks" onclick="openTab(event, 'Restore')" id="defaultOpen">Restore</button>
-            <button class="tablinks" onclick="openTab(event, 'Slice')">Slice</button>
-            <button class="tablinks" onclick="openTab(event, 'Chord')">Chords</button>
-            <button class="tablinks" onclick="openTab(event, 'SynthPresetInspector')">Macros</button>
-            <button class="tablinks" onclick="openTab(event, 'Reverse')">Reverse</button>
-            <button class="tablinks" onclick="openTab(event, 'MidiUpload')">MIDI Upload</button>
-        </div>
-
-        <div id="Restore" class="tabcontent">
-            <!-- The Restore Set form content will be loaded here dynamically -->
-        </div>
-
-        <div id="Slice" class="tabcontent">
-            <!-- The Slice Kit form content will be loaded here dynamically -->
-        </div>
-
-        <div id="Chord" class="tabcontent">
-            <!-- The Chord Kit form content will be loaded here dynamically -->
-        </div>
-
-        <div id="SynthPresetInspector" class="tabcontent">
-            <!-- Synth Preset Inspector functionality will be added here later -->
-        </div>
-
-        <div id="Reverse" class="tabcontent">
-            <!-- The Reverse WAV form content will be loaded here dynamically -->
-        </div>
-
-        <div id="MidiUpload" class="tabcontent">
-            <!-- The MIDI Upload form content will be loaded here dynamically -->
-        </div>
-
-<small id="footer">More information on <a href="http://github.com/charlesvestal/extending-move">GitHub</a>. <a href="mailto:charles@charles.pizza">Feedback welcome!</a></small>
-
-    </body>
-</html>
