--- templates/midi_upload.html
@@ -1,55 +1,0 @@
-<h2>MIDI Upload</h2>
-<div id="result-message">{message_html}</div>
-
-<div  >
-    <div  >
-        <h3>Upload MIDI File</h3>
-        <p>Upload a MIDI file to generate an Ableton Live set with the notes from the file.</p>
-        
-        <form method="post" action="/midi-upload" enctype="multipart/form-data">
-            <input type="hidden" name="action" value="upload_midi" />
-            
-            <div class="form-group">
-                <label for="midi_type">MIDI Type:</label>
-                <select name="midi_type" id="midi_type" required>
-                    <option value="melodic" selected>Melodic MIDI</option>
-                    <option value="drum">Drum MIDI (808)</option>
-                </select>
-            </div>
-            
-            <div class="form-group">
-                <label for="midi_file">MIDI File:</label>
-                <input type="file" name="midi_file" id="midi_file" accept=".mid,.midi" required />
-                <small>Supported formats: .mid, .midi</small>
-            </div>
-            
-            <div class="form-group">
-                <label for="set_name">Set Name:</label>
-                <input type="text" name="set_name" id="set_name" placeholder="My MIDI Set" required />
-            </div>
-
-            <div class="form-group">
-                <label for="tempo">Tempo (BPM):</label>
-                <input type="number" name="tempo" id="tempo" min="20" max="300" placeholder="Auto-detect from MIDI" />
-                <small>Leave blank to use tempo from MIDI file (or default 120 BPM)</small>
-            </div>
-
-            <div class="form-group">
-                <label for="pad_index">Device Pad:</label>
-                <select name="pad_index" id="pad_index" required>
-                    {{ pad_options }}
-                </select>
-            </div>
-            
-            <div class="form-group">
-                <label for="pad_color">Pad Color:</label>
-                <select name="pad_color" id="pad_color" required>
-                    {{ pad_color_options }}
-                </select>
-            </div>
-            
-            <button type="submit" >Generate Set from MIDI</button>
-        </form>
-    </div>
-</div>
-
