--- templates/refresh.html
@@ -1,6 +1,0 @@
-<h2>Manually Refresh the Move's Library</h2>
-{message_html}
-<form method="post">
-    <input type="hidden" name="action" value="refresh_library"/>
-    <input type="submit" value="Refresh Library"/>
-</form>
