--- templates/restore.html
@@ -1,18 +1,0 @@
-<h2>Restore Move Set (.ablbundle)</h2>
-    <form action="/restore" method="post" enctype="multipart/form-data">
-        <label for="ablbundle">Select .ablbundle file:</label>
-        <input type="file" name="ablbundle" accept=".ablbundle" required>
-        <br>
-        <label for="mset_index">Restore to pad:</label>
-        <select name="mset_index">
-            {{ options }}
-        </select>
-        <br>
-        <label for="mset_color">Pad color (1-26):</label>
-        <input type="number" name="mset_color" min="1" max="26" value="1" required>
-        <br>
-        <input type="hidden" name="action" value="restore_ablbundle">
-        <button type="submit">Upload & Restore</button>
-    </form>
-    <div id="result-message">{message_html}</div>
-    <img src="/static/img/move_pads.png" id="padids" style="width: 30%;">
