--- templates/synth_preset_inspector.html
@@ -1,181 +1,0 @@
-<h2>Synth Macros (Drift)</h2>
-<p><em>Note: Deleting and adding macro assignments will immediately alter your preset files and can result in broken presets. Back them up! Ranges are optional, and not all parameters take them (i.e. Voice1_Filter1_Type). Experiement!</em></p>
-
-{message_html}
-
-<form method="POST" action="/synth-preset-inspector" id="preset-form">
-    <input type="hidden" name="action" id="action-input" value="select_preset">
-    <input type="hidden" name="param_path" id="param-path-input" value="">
-    <input type="hidden" name="param_name" id="param-name-input" value="">
-    <input type="hidden" name="macro_index" id="macro-index-input" value="">
-    <div class="preset-selection">
-        <select name="preset_select" id="preset_select">
-            {{ options }}
-        </select>
-    </div>
-    
-    <div class="macro-display">
-        {{ macros_html }}
-    </div>
-</form>
-
-<!-- The dropdown change event is handled by main.js using AJAX -->
-
-<style>
-    .preset-selection {
-        margin-bottom: 20px;
-        padding: 15px;
-        background-color: #f5f5f5;
-        border-radius: 5px;
-        display: flex;
-        align-items: center;
-        gap: 10px;
-    }
-    
-    .preset-selection select {
-        flex-grow: 1;
-        padding: 8px;
-        border: 1px solid #ccc;
-        border-radius: 4px;
-    }
-    
-    .macros-container {
-        margin-top: 20px;
-    }
-    
-    .macro-item {
-        margin-bottom: 20px;
-        padding: 15px;
-        border: 1px solid #ddd;
-        border-radius: 5px;
-        background-color: #f9f9f9;
-    }
-    
-    .macro-header {
-        margin-top: 0;
-        color: #333;
-        font-size: 1.17em;
-        font-weight: bold;
-        margin-bottom: 15px;
-        display: flex;
-        align-items: center;
-    }
-    
-    .macro-name-input {
-        padding: 5px;
-        border: 1px solid #ccc;
-        border-radius: 3px;
-        font-size: 1em;
-        width: 60%;
-        margin-left: 10px;
-    }
-    
-    .parameter-mapping {
-        margin: 15px 0;
-        padding: 10px;
-        background-color: #f0f0f0;
-        border-radius: 4px;
-    }
-    
-    .parameter-mapping select {
-        display: block;
-        width: 100%;
-        padding: 8px;
-        margin: 8px 0 15px;
-        border: 1px solid #ccc;
-        border-radius: 4px;
-    }
-    
-    .parameter-controls {
-        display: flex;
-        align-items: center;
-        gap: 15px;
-        margin-bottom: 10px;
-    }
-    
-    .range-inputs-inline {
-        display: flex;
-        gap: 10px;
-    }
-    
-    .range-inputs-inline label {
-        display: flex;
-        align-items: center;
-        font-size: 0.9em;
-        color: #555;
-    }
-    
-    .range-inputs-inline input {
-        padding: 6px;
-        border: 1px solid #ccc;
-        border-radius: 3px;
-        width: 80px;
-        margin-left: 5px;
-    }
-    
-    .current-mappings {
-        margin-top: 15px;
-        padding-top: 10px;
-        border-top: 1px dashed #ccc;
-    }
-    
-    .current-mappings h4 {
-        margin-top: 0;
-        font-size: 1em;
-        color: #555;
-    }
-    
-    .parameters-list {
-        margin-top: 10px;
-        padding-left: 20px;
-    }
-    
-    .parameters-list li {
-        margin-bottom: 5px;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-    }
-    
-    .parameter-info {
-        flex-grow: 1;
-    }
-    
-    .delete-mapping-btn {
-        background-color: #f44336;
-        color: white;
-        border: none;
-        border-radius: 3px;
-        padding: 3px 8px;
-        font-size: 0.8em;
-        cursor: pointer;
-        margin-left: 10px;
-    }
-    
-    .delete-mapping-btn:hover {
-        background-color: #d32f2f;
-    }
-    
-    .add-mapping-btn {
-        background-color: #4CAF50;
-        color: white;
-        border: none;
-        border-radius: 3px;
-        padding: 5px 10px;
-        font-size: 0.9em;
-        cursor: pointer;
-        margin-left: 10px;
-    }
-    
-    .add-mapping-btn:hover {
-        background-color: #45a049;
-    }
-    
-    .save-name-btn {
-        margin-left: 10px;
-    }
-    
-    .save-name-btn:hover {
-        background-color: #0b7dda;
-    }
-</style>
